+++ d3d12game_xdk_cppwinrt_dr/DeviceResources.h
@@ -64,7 +64,9 @@ namespace DX
 
         CD3DX12_CPU_DESCRIPTOR_HANDLE GetGameDVRRenderTargetView() const
         {
-            return CD3DX12_CPU_DESCRIPTOR_HANDLE(m_rtvDescriptorHeap->GetCPUDescriptorHandleForHeapStart(), m_backBufferCount + m_backBufferIndex, m_rtvDescriptorSize);
+            return CD3DX12_CPU_DESCRIPTOR_HANDLE(
+                m_rtvDescriptorHeap->GetCPUDescriptorHandleForHeapStart(),
+                static_cast<INT>(m_backBufferCount + m_backBufferIndex), m_rtvDescriptorSize);
         }
 
     private:
